autonomy_gatekeeper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autonomy-gatekeeper"
version = "0.1.0"
description = "AI governance agent that decides whether to ACT, HOLD, or ESCALATE."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-openai>=0.2,<0.3",
    "langgraph>=0.2,<0.3",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0,<14.0",
    "click>=8.0,<9.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.11,<2.0",
    "pre-commit>=3.0,<4.0",
    "types-PyYAML>=6.0,<7.0",
]

[project.scripts]
autonomy-gatekeeper = "autonomy_gatekeeper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/autonomy_gatekeeper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
